+++ 2019/08/template/heiseirb_api_client/lib/heiseirb_api_client/example.rb
@@ -1,11 +1,7 @@
 require "heiseirb_api_client/core"
-require "securerandom"
-require "heiseirb_api_client/user"
-require 'json'
 
 module HeiseirbApiClient
   class Example < Core
-
     def initialize
       @conn = Faraday.new(url: URL)  do |config|
         config.request :json
@@ -23,14 +19,6 @@ module HeiseirbApiClient
       response.body
     end
 
-    def get_event(event_id)
-      response = @conn.get do |req|
-        req.url "/client/events/#{event_id}"
-        req.headers = HEADRS
-      end
-      response.body
-    end
-
     def get_users(page = 1)
       response = @conn.get do |req|
         req.url '/client/users'
@@ -38,46 +26,6 @@ module HeiseirbApiClient
         req.params['page'] = page
       end
       response.body
-    end
-
-    def get_user(user_id)
-      response = @conn.get do |req|
-        req.url "/client/users/#{user_id}"
-        req.headers = HEADRS
-      end
-      User.new(response.body['data']['attributes'])
-    end
-
-    def update_user(user_id, params: {})
-      response = @conn.patch do |req|
-        req.url "/client/users/#{user_id}"
-        req.headers = HEADRS
-        req.body = {
-          user: {
-            name: "hogehgoe",
-            password: "password",
-            password_confirmation: "password"
-          }
-        }
-      end
-      puts response
-      # User.new(response.body['data']['attributes'])
-      response.body
-    end
-
-    def bulk_create_users(num)
-      num.times do
-        Thread.new do
-          num.times do |i|
-            randam = SecureRandom.hex(i)
-            name = "tashiro_#{randam}"
-            email = "example-#{randam}@gmail.com"
-            role = "owner"
-            password = "hogehoge#{randam}"
-            create_user(name, email, role, password)
-          end
-        end
-      end
     end
 
     def create_user(name, email, role, password)
