+++ 2019/08/template/heiseirb_api_client/lib/heiseirb_api_client/models/user.rb
